Added addCode() function to ConnectionMongoDB class.
----------------------------------------------------

It allows the insertion of a new code as soon as a user uses it. The new code will have the field "code" different from the one just used but will have the usual amount in the field "credit" so as to maintain the collection balanced from the point of view of the field "credit".